# docs/api/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Laravel API",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "https://settlane.tech"
        }
    ],
    "paths": {
        "/api/admin/dashboard": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "DashboardController",
                "operationId": "get_api_admin_dashboard",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "SessionAuth": []
                    }
                ]
            }
        },
        "/api/admin/invoices": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "InvoiceController::index",
                "operationId": "get_api_admin_invoices",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "SessionAuth": []
                    }
                ]
            }
        },
        "/api/admin/invoices/{invoice}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "InvoiceController::show",
                "operationId": "get_api_admin_invoices__invoice_",
                "parameters": [
                    {
                        "name": "invoice",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "SessionAuth": []
                    }
                ]
            }
        },
        "/api/admin/invoices/{invoice}/refresh": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "InvoiceController::refresh",
                "operationId": "post_api_admin_invoices__invoice__refresh",
                "parameters": [
                    {
                        "name": "invoice",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "SessionAuth": []
                    }
                ]
            }
        },
        "/api/admin/merchant-api-keys": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "MerchantApiKeyController::index",
                "operationId": "get_api_admin_merchant_api_keys",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "SessionAuth": []
                    }
                ]
            }
        },
        "/api/admin/merchant-api-keys/{apiKey}/revoke": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "MerchantApiKeyController::revoke",
                "operationId": "post_api_admin_merchant_api_keys__apiKey__revoke",
                "parameters": [
                    {
                        "name": "apiKey",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "SessionAuth": []
                    }
                ]
            }
        },
        "/api/admin/merchant-users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "MerchantUserController::index",
                "operationId": "get_api_admin_merchant_users",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "SessionAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "MerchantUserController::store",
                "operationId": "post_api_admin_merchant_users",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "merchant_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "role_id": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "merchant_id",
                                    "name",
                                    "email",
                                    "password",
                                    "role_id",
                                    "status"
                                ]
                            }
                        }
                    }
                },
                "security": [
                    {
                        "SessionAuth": []
                    }
                ]
            }
        },
        "/api/admin/merchant-users/{merchantUser}/role": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "MerchantUserController::updateRole",
                "operationId": "patch_api_admin_merchant_users__merchantUser__role",
                "parameters": [
                    {
                        "name": "merchantUser",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "role_id": {
                                        "type": "integer"
                                    }
                                },
                                "required": [
                                    "role_id"
                                ]
                            }
                        }
                    }
                },
                "security": [
                    {
                        "SessionAuth": []
                    }
                ]
            }
        },
        "/api/admin/merchant-users/{merchantUser}/status": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "MerchantUserController::updateStatus",
                "operationId": "patch_api_admin_merchant_users__merchantUser__status",
                "parameters": [
                    {
                        "name": "merchantUser",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "status"
                                ]
                            }
                        }
                    }
                },
                "security": [
                    {
                        "SessionAuth": []
                    }
                ]
            }
        },
        "/api/admin/merchants": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "MerchantController::index",
                "operationId": "get_api_admin_merchants",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "SessionAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "MerchantController::store",
                "operationId": "post_api_admin_merchants",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "fee_percent": {
                                        "type": "number"
                                    },
                                    "webhook_url": {
                                        "type": "string"
                                    },
                                    "webhook_secret": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "name"
                                ]
                            }
                        }
                    }
                },
                "security": [
                    {
                        "SessionAuth": []
                    }
                ]
            }
        },
        "/api/admin/merchants/{merchant}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "MerchantController::show",
                "operationId": "get_api_admin_merchants__merchant_",
                "parameters": [
                    {
                        "name": "merchant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "SessionAuth": []
                    }
                ]
            }
        },
        "/api/admin/merchants/{merchant}/status": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "MerchantController::updateStatus",
                "operationId": "patch_api_admin_merchants__merchant__status",
                "parameters": [
                    {
                        "name": "merchant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "status"
                                ]
                            }
                        }
                    }
                },
                "security": [
                    {
                        "SessionAuth": []
                    }
                ]
            }
        },
        "/api/admin/merchants/{merchant}/wallets": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "MerchantWalletController::index",
                "operationId": "get_api_admin_merchants__merchant__wallets",
                "parameters": [
                    {
                        "name": "merchant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "SessionAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "MerchantWalletController::store",
                "operationId": "post_api_admin_merchants__merchant__wallets",
                "parameters": [
                    {
                        "name": "merchant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "SessionAuth": []
                    }
                ]
            }
        },
        "/api/admin/merchants/{merchant}/wallets/{wallet}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "MerchantWalletController::update",
                "operationId": "put_api_admin_merchants__merchant__wallets__wallet_",
                "parameters": [
                    {
                        "name": "merchant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "wallet",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "SessionAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "MerchantWalletController::destroy",
                "operationId": "delete_api_admin_merchants__merchant__wallets__wallet_",
                "parameters": [
                    {
                        "name": "merchant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "wallet",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "SessionAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "MerchantApiKey": {
                "type": "http",
                "scheme": "bearer"
            },
            "SessionAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "laravel-session"
            }
        }
    }
}